Adding 64-bit int support when rendering. Changes include:
  1) Adding new Int64 and UInt64 types
  2) Improved how OpenGL extensions are queried across multi-gpus and
     handled in shaders. This was necessary as 64-bit int support is
     only available in extensions.
  3) Hit-testing was improved to handle 64-bit row ids. Neither opengl
     nor any extensions support 64-bit int textures, so to support this
     required adding two 32-bit int textures to pack the 64-bit row
     id into. Hit testing will now work with tables with rows > 4.3 billion.
  4) Changed how a mark's render properties manage their internal value
     so that it can be used directly in scales.
  5) Fixed some various bugs, such as reintroducing non-rendering of
     empty rows, and fixed some small update issues.

+++ QueryRenderer/Marks/PointMark.h
@@ -29,7 +29,7 @@ class PointMark : public BaseMark {
   RenderProperty<float> size;
   rapidjson::Pointer _sizeJsonPath;
 
-  RenderProperty<unsigned int> id;
+  RenderProperty<uint64_t> id;
   rapidjson::Pointer _idJsonPath;
 
   ColorRenderProperty fillColor;
